requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv
duckduckgo-search
ddgs
groq
supabase
sentence-transformers